Add exists check before unlink

+++ scripts/failed-tests.js
@@ -80,7 +80,7 @@ class FailedTestsReporter extends Mocha.reporters.Base {
             const failed = Array.from(failingTests).join(os.EOL);
             fs.writeFile(file, failed, "utf8", done);
         }
-        else if (!keepFailed) {
+        else if (!keepFailed && fs.existsSync(file)) {
             fs.unlink(file, done);
         }
         else {
